Formatting Src dir
Fixes #8
